backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
groq==0.11.0
chromadb==0.5.15
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
pypdf==4.3.1
sentence-transformers==3.1.1
python-multipart==0.0.12
PyJWT==2.9.0